// maeri/src/app/models/product.model.ts
export class Product {
  adminId: string;
  categoryName: string;
  name: string;
  capacity: string;
  productImage: string;
  description: string;
  purchasingPrice: number;
  sellingPrice: number;
  superMarketPrice: number;
  randomPrice: number;
  quantityMin: number;
  sizeUnit: number;
  unitName: string;
  produceBy: string;
  packPrice: number;
  packSize: number;
  ristourne: number;
  bonus: number;
  categoryId: string;
  categoryIndicator: string;
  created: string;
  dateupdate: string;
  desabled: false;
  disablemanufactured: true;
  display: true;
  filename: string;
  isChecked: boolean;
  idList?: string[];
  newstock: number;
  oldstock: number;
  originalname: string;
  productType: string;
  quantity: number;
  quantityItems: number;
  quantityStore: number;
  resourceList: [];
  retailerPrice: 0;
  storeId: string;
  superCategory: string;
  tabitem: [];
  url: string;
  __v: 0;
  _id: string;
  toRemove: number;
  quantityRandom: number;
  quantityToAlert: number;
  reserve?: number = 0;
  dispo?: number = 0;
  changes?: boolean = false;
  grammage: number = 0;
  masse: number = 0;
  bottle_full?: number = 0;
  bottle_total?: number = 0;
  bottle_empty?: number = 0;
  constructor() {}
}